Make "Best Girl" voting db
Simple file on disk because why go complex when simple work fine?

Going to let people use channel points to vote for which of the
characters available are the best. I'll put Rei and Asuka up first
obviously so that Gendo can complain that they've been at this for 20+
years.